Move Node into its own files
