+++ core/silkworm/trie/hash_builder.cpp
@@ -28,32 +28,7 @@
 #include <silkworm/common/util.hpp>
 #include <silkworm/rlp/encode.hpp>
 
-static void assert_subset(uint16_t sub, uint16_t sup) {
-    auto intersection{sub & sup};
-    assert(intersection == sub);
-    (void)intersection;
-}
-
 namespace silkworm::trie {
-
-Node::Node(uint16_t state_mask, uint16_t tree_mask, uint16_t hash_mask, std::vector<evmc::bytes32> hashes,
-           std::optional<evmc::bytes32> root_hash)
-    : state_mask_{state_mask},
-      tree_mask_{tree_mask},
-      hash_mask_{hash_mask},
-      hashes_{std::move(hashes)},
-      root_hash_{std::move(root_hash)} {
-    assert_subset(tree_mask_, state_mask_);
-    assert_subset(hash_mask_, state_mask_);
-    assert(std::bitset<16>(hash_mask_).count() == hashes_.size());
-}
-
-void Node::set_root_hash(std::optional<evmc::bytes32> root_hash) { root_hash_ = std::move(root_hash); }
-
-bool operator==(const Node& a, const Node& b) {
-    return a.state_mask() == b.state_mask() && a.tree_mask() == b.tree_mask() && a.hash_mask() == b.hash_mask() &&
-           a.hashes() == b.hashes() && a.root_hash() == b.root_hash();
-}
 
 Bytes unpack_nibbles(ByteView packed) {
     Bytes out(2 * packed.length(), '\0');
